test: update testing
